+++ ospd/scan.py
@@ -356,16 +356,12 @@ class ScanCollection:
             self.scans_table[scan_id]['end_time'] = int(time.time())
 
     def get_status(self, scan_id: str) -> ScanStatus:
-<<<<<<< HEAD
         """Get scan_id scans's status."""
-=======
-        """ Get scan_id scans's status."""
         status = None
         try:
             status = self.scans_table[scan_id].get('status')
         except KeyError:
-            LOGGER.error("Scan ID %s not found", scan_id)
->>>>>>> a682e99 (Fix: Fix get_status (#471))
+            logger.error("Scan ID %s not found", scan_id)
 
         return status
 
@@ -434,7 +430,7 @@ class ScanCollection:
             Count of excluded host.
         """
         exc_hosts_list = target_str_to_list(self.get_exclude_hosts(scan_id))
-        LOGGER.debug(
+        logger.debug(
             '%s: Excluded Hosts: %s',
             scan_id,
             pformat(exc_hosts_list),
@@ -443,7 +439,7 @@ class ScanCollection:
         finished_hosts_list = target_str_to_list(
             self.get_finished_hosts(scan_id)
         )
-        LOGGER.debug(
+        logger.debug(
             '%s: Finished Hosts: %s',
             scan_id,
             pformat(finished_hosts_list),